Move MediaTranslationAdmin to a more sensible spot and dont hook up ResourceBase for translation directly (only do it with Layer, Map, Document etc)
