+++ geonode/documents/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 from geonode.documents.models import Document
-from geonode.utils import MediaTranslationAdmin
+from geonode.base.admin import MediaTranslationAdmin
 
 class DocumentAdmin(MediaTranslationAdmin):
     list_display = ('id', 'title', 'date', 'category')
