+++ geonode/utils.py
@@ -33,7 +33,6 @@ from django.shortcuts import get_object_or_404
 from django.utils import simplejson as json
 from django.http import HttpResponse
 from geonode.security.enumerations import AUTHENTICATED_USERS, ANONYMOUS_USERS, INVALID_PERMISSION_MESSAGE
-from modeltranslation.admin import TranslationAdmin
 from urlparse import urlsplit
 
 DEFAULT_TITLE=""
@@ -580,14 +579,3 @@ def json_response(body=None, errors=None, redirect_to=None, exception=None,
        body = json.dumps(body)
    return HttpResponse(body, content_type=content_type, status=status)
 
-
-class MediaTranslationAdmin(TranslationAdmin):
-    class Media: 
-        js = (
-            'http://ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js',
-            'http://ajax.googleapis.com/ajax/libs/jqueryui/1.10.2/jquery-ui.min.js',
-            'modeltranslation/js/tabbed_translation_fields.js',
-        )
-        css = {
-            'screen': ('modeltranslation/css/tabbed_translation_fields.css',),
-        }
